@@ -3,27 +3,25 @@
     "version": "1.0.0",
     "main": "index.js",
     "scripts": {
-        "test": "jest"
+        "test": "jest ./test"
     },
-    "author": "cuc_ygh@163.com",
-    "license": "Apache-2.0",
     "dependencies": {
-        "@babel/core": "7.1.2",
-        "@babel/preset-env": "7.1.6",
-        "babel-core": "7.0.0-bridge.0",
-        "babel-jest": "23.6.0",
+        "@babel/core": "7.6.4",
+        "@babel/preset-env": "7.6.3",
+        "babel-core": "6.26.3",
+        "babel-jest": "24.9.0",
         "babel-preset-env": "1.7.0",
-        "eslint-plugin-jest": "22.0.0",
-        "jest": "23.6.0",
-        "regenerator-runtime": "0.12.1"
+        "eslint-plugin-jest": "22.17.0",
+        "jest": "24.9.0",
+        "regenerator-runtime": "0.13.3"
     },
     "devDependencies": {
-        "eslint": "5.9.0",
-        "eslint-config-standard": "12.0.0",
-        "eslint-plugin-import": "2.14.0",
-        "eslint-plugin-node": "8.0.0",
-        "eslint-plugin-promise": "4.0.1",
-        "eslint-plugin-standard": "4.0.0"
+        "eslint": "6.5.1",
+        "eslint-config-standard": "14.1.0",
+        "eslint-plugin-import": "2.18.2",
+        "eslint-plugin-node": "10.0.0",
+        "eslint-plugin-promise": "4.2.1",
+        "eslint-plugin-standard": "4.0.1"
     },
     "directories": {
         "test": "test"
