@@ -1,3 +1,8 @@
 module.exports = {
-    "extends": ["standard","plugin:jest/recommended"]
+    "extends": ["standard", "plugin:jest/recommended"],
+    rules: {
+        'indent': [4, 4, {
+            'SwitchCase': 2 // 针对switch case的缩进
+        }],
+    }
 };
