@@ -1,11 +1,7 @@
-// import revertByWorld from "../../code/string/lesson1";
+import revertByWorld from "../../code/string/lesson1";
 
-// test("revertByWorld:Let's take LeetCode contest", () => {
-//     expect(revertByWorld("Let's take LeetCode contest")).toBe("s'teL ekat edoCteeL tsetnoc");
-// });
-
-import sum from '../../test/string/lesson1.test'
-
-test('adds 1 + 2 to equal 3', () => {
-    expect(sum(1, 2)).toBe(3);
+test("revertByWorld:Let's take LeetCode contest", () => {
+    expect(revertByWorld("Let's take LeetCode contest")).toBe(
+        "s'teL ekat edoCteeL tsetnoc"
+    );
 });
